# Add extensive error output for get_offset