Might help diagnose https://github.com/julia-vscode/LanguageServer.jl/issues/501.

I tried it once with a fake error, it seems to work. Not the nicest output, but hopefully we'll get what we need. Do you think this is sufficient?