@ZacLN 2020-01-28
How much overhead is there for wrapping stuff in a try block? This func gets called alot... If the overheads significant (rough guess) then can we wrap the whole function in a try block instead? (i.e. rename the existing `get_offst` -> `_get_offset` and have get_index call _get_index in a try block)