@davidanthoff 2020-01-28
Good point :) Here are some benchmarks:

Without try catch
```
julia> @benchmark LanguageServer.get_offset(d, 37, 8)
BenchmarkTools.Trial:
  memory estimate:  64 bytes
  allocs estimate:  1
  --------------
  minimum time:     45.501 ns (0.00% GC)
  median time:      50.960 ns (0.00% GC)
  mean time:        58.058 ns (1.94% GC)
  maximum time:     1.134 μs (93.05% GC)
  --------------
  samples:          10000
  evals/sample:     989
```
With my original implementation:
```
julia> @benchmark LanguageServer.get_offset(d, 37, 8)
BenchmarkTools.Trial:
  memory estimate:  128 bytes
  allocs estimate:  2
  --------------
  minimum time:     167.285 ns (0.00% GC)
  median time:      186.359 ns (0.00% GC)
  mean time:        201.089 ns (1.15% GC)
  maximum time:     1.669 μs (76.86% GC)
  --------------
  samples:          10000
  evals/sample:     755
```
With the new implementation that I just pushed:
```
julia> @benchmark LanguageServer.get_offset(d, 37, 8)
BenchmarkTools.Trial:
  memory estimate:  128 bytes
  allocs estimate:  2
  --------------
  minimum time:     69.928 ns (0.00% GC)
  median time:      79.609 ns (0.00% GC)
  mean time:        95.573 ns (2.44% GC)
  maximum time:     1.142 μs (87.24% GC)
  --------------
  samples:          10000
  evals/sample:     971
```
So still slower, but maybe ok? I think it would be good if we had access to the `position(io)` value, but we only get that if we start the try block after the `IOBuffer` is constructed.