@ZacLN 2020-01-28
That looks fine to me (I had no ballpark idea for the slow down)