@ZacLN 2020-01-29
Sorry, I missed the part dumping text. Can we obscure the text so that we're not getting sight of users code? Not particularly elegant or efficient but would the following work?
```julia
function obscure_text(s)
    i = 1
    io = IOBuffer()
    while i <= sizeof(s)
        di = nextind(s, i) - i
        if di == 1
            if s[i] in ('\n', '\r')
                write(io, s[i])
            else
                write(io, "a")
            end
        elseif di == 2
            write(io, "α")
        elseif di == 3
            write(io, "—")
        else
            write(io, s[i])
        end
        i += di
    end
    s1 = String(take!(io))
end
```