@davidanthoff 2020-01-29
Yes, that would probably work.

I think hunting that bug down will be tricky, we only had three crash reports for it from one user... So I'm not even sure we'll get more reports of this from our beta testers... Can you open a PR with this obscure stuff? I don't have much more time left today :)